requirements.txt
# ============================================================
# GeoExplorer AI — Python Dependencies
# ============================================================
# Core geo / spatial
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0
rasterio>=1.3.0
fiona>=1.9.0
srtm.py>=0.3.6

# Data
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
tqdm>=4.65.0

# ML
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0
shap>=0.43.0
imbalanced-learn>=0.11.0

# Visualisation
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
folium>=0.15.0
pydeck>=0.8.0

# Streamlit frontend
streamlit>=1.28.0
streamlit-folium>=0.15.0
streamlit-extras>=0.3.6

# LLM / AI
huggingface_hub>=0.19.0
httpx>=0.25.0

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0
rich>=13.6.0
click>=8.1.0
loguru>=0.7.0
